Problem: tests.py becomes very long
Solution: move to directory
starts to chop into pieces
