+++ src/zc/buildout/tests/bootstrap_cl_settings.test
