+++ src/zc/buildout/tests/easy_install.txt
