+++ src/zc/buildout/tests/repeatable.txt
